The inputs builder separated for every operation
