+++ cs-hashicorp-terraform/src/test/java/io/cloudslang/content/hashicorp/terraform/services/RunImplTest.java
@@ -17,6 +17,7 @@ package io.cloudslang.content.hashicorp.terraform.services;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import io.cloudslang.content.hashicorp.terraform.entities.CreateRunInputs;
+import io.cloudslang.content.hashicorp.terraform.entities.TerraformCommonInputs;
 import io.cloudslang.content.hashicorp.terraform.utils.Inputs;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -30,10 +31,9 @@ public class RunImplTest {
     private final String EXPECTED_CREATE_RUN_BODY="{\"data\":{\"attributes\":{\"is-Destroy\":\"false\",\"message\":\"test\"},\"type\":\"runs\",\"relationships\":{\"workspace\":{\"data\":{\"type\":\"workspaces\",\"id\":\"test-123\"}}}}}";
     private final CreateRunInputs invalidCreateRunInputs=CreateRunInputs.builder()
             .workspaceId("")
-            .workspaceName("")
             .runMessage("")
             .isDestroy("")
-            .commonInputs(Inputs.builder()
+            .commonInputs(TerraformCommonInputs.builder()
                     .authToken("")
                     .requestBody("")
                     .proxyHost("")
